wallet/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
